requirements.txt
fastapi
uvicorn
python-dotenv
numpy
opencv-python
mediapipe
torch
scikit-learn
fastapi
uvicorn
requests
scipy
pydantic
pandas